chore: 평가·TTS·PII·삼성헬스·개발일지 작업 일괄 커밋
- evaluation: 삼성헬스 export 파서·헬스 어댑터 검증, 회복점수 재설계 문서
- pii: 위기감지 PII 비식별(redact) + safety/emotion/message 호출부 연결
- tts: qwen3 합성·서버 보완, TTS URL 자동등록·트러블슈팅 핸드오프 문서
- docs: 개발일지(정환주) + 통합 일지, AGENTS.md

+++ ai/tts/qwen3_emotion.py
@@ -100,7 +100,7 @@ def _load_model():
         device, dtype, attn = "cpu", torch.float32, None
     else:
         print(f"[gpu] {torch.cuda.get_device_name(0)}")
-        device, dtype, attn = "cuda:0", torch.bfloat16, None  # flash-attn 미설치 → 기본
+        device, dtype, attn = "cuda:0", torch.float16, None
 
     kwargs = dict(device_map=device, dtype=dtype)
     if attn:
